+++ ProjectForOPI/Events.h
@@ -5,6 +5,7 @@
 #include "Physics.h"
 #include "FightSystem.h"
 #include "Vars.h"
+#include "Compontent.h"
 
 using namespace CE;
 using namespace std;
@@ -16,16 +17,19 @@ namespace Game {
 
 		cout << "Event is work!!!" << endl;
 
+		
 		system("pause");
+		system("cls");
 	}
 	
 	void EventF() {
 		system("cls");
 
-		cout << Enemys[0].name << endl << Enemys[0].health << endl << Enemys[0].armor << endl << Enemys[0].attack << endl 
-			<< Enemys[0].discription << endl;
+		cout << Enemys[1].name << endl << Enemys[1].health << endl << Enemys[1].armor << endl << Enemys[1].attack << endl 
+			<< Enemys[1].discription << endl;
 
 		system("pause");
+		system("cls");
 	}
 
 	void EventUnit() {
