+++ ProjectForOPI/ProjectForOPI.vcxproj
@@ -22,8 +22,8 @@
     <ClCompile Include="Source.cpp" />
   </ItemGroup>
   <ItemGroup>
+    <ClInclude Include="Compontent.h" />
     <ClInclude Include="DefaultMath.h" />
-    <ClInclude Include="Enemy.h" />
     <ClInclude Include="Events.h" />
     <ClInclude Include="FightSystem.h" />
     <ClInclude Include="GameLogic.h" />
